画像データの送受信が可能に
