+++ unity-sample-environment/Assets/Scripts/AIServer.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using WebSocketSharp;
 using WebSocketSharp.Server;
 using WebSocketSharp.Net;
+using System.Threading;
 using MsgPack;
 
 namespace MLPlayer
@@ -11,61 +12,56 @@ namespace MLPlayer
 	public class AIServer : MonoBehaviour
 	{
 		private WebSocketServer wssv;
-		                                           //*******************************************************
-		private Queue<byte[]> agentMessageQueue;  //ここのagentMessageQueueのデータを下にあるクラスのCommuncationGymに持っていきたい
-		private Queue<byte[]> aiMessageQueue;     //*******************************************************
+		                                          
+		public Queue<byte[]> agentMessageQueue;
+		private Queue<byte[]> aiMessageQueue;
+		private Mutex mutAgent;
 		public Agent agent;
 		private MsgPack.CompiledPacker packer;
 
-		public AIServer (Agent _agent) {
-			packer = new MsgPack.CompiledPacker();
-			agentMessageQueue = new Queue<byte[]>();
-			aiMessageQueue = new Queue<byte[]>();
+		public AIServer (Agent _agent)
+		{
 			agent = _agent;
+			mutAgent = new Mutex ();
+			packer = new MsgPack.CompiledPacker ();
+			agentMessageQueue = new Queue<byte[]> ();
+			aiMessageQueue = new Queue<byte[]> ();
 		}
-			
+
 		public class CommunicationGym : WebSocketBehavior
 		{
 			public Agent agent { set; get; }
-			public State CheckMSG = new State ();  
- 			MsgPack.BoxingPacker packer = new MsgPack.BoxingPacker ();
-
-			//*******************************************
-			//private Queue<byte[]> agentMessageQueue; こうすると当然別のキューが作成されてしまう...
-			//*******************************************
-
+			MsgPack.BoxingPacker packer = new MsgPack.BoxingPacker ();
+			private bool SendFlag=false;
 
 			protected override void OnMessage (MessageEventArgs e)
 			{
-				//receive message from GYM
-				agent.action.Set((Dictionary<System.Object,System.Object>)packer.Unpack (e.RawData));
-				Debug.Log ("Rotate="+agent.action.rotate+" Forword="+agent.action.forward+" Jump="+agent.action.jump);
+				//receive message 
+				agent.action.Set ((Dictionary<System.Object,System.Object>)packer.Unpack (e.RawData));
+				SceneController.received.Set ();
+				Debug.Log ("Rotate=" + agent.action.rotate + " Forword=" + agent.action.forward + " Jump=" + agent.action.jump);
 
+				SendFlag = false;
 
-				//byte[] data = PopAgentState();
-				//Send (data);
-
-
-				SendMessage (CheckMSG);
+				//send state data 
+				while (SendFlag == false) {
+					if (SceneController.server.agentMessageQueue.Count > 0) {
+						byte[] data = SceneController.server.PopAgentState ();
+						Send (data);
+						SendFlag = true;
+					}
+				}
 			}
-				
-			public void SendMessage(State s){
-				//Send state datas to GYM
-				var packer2 = new MsgPack.CompiledPacker ();
-				byte[] msg = packer2.Pack (s);
-				Send (msg);
-			}
-				
+
 			protected override void OnOpen ()
 			{
 				Debug.Log ("Socket Open");
 			}
-
-		
 		}
 
-		CommunicationGym instantiate() {
-			CommunicationGym service = new CommunicationGym();
+		CommunicationGym instantiate ()
+		{
+			CommunicationGym service = new CommunicationGym ();
 			service.agent = agent;
 			return service;
 		}
@@ -93,20 +89,27 @@ namespace MLPlayer
 			throw new System.NotImplementedException ();
 		}
 
-		public void PushAgentState(State s) {
-			byte[] msg = packer.Pack(s);     //*********************************
-			agentMessageQueue.Enqueue(msg);  //キューには画像データが入っていることを確認済み
-		}                          
+		public void PushAgentState (State s)
+		{
+			byte[] msg = packer.Pack (s);  
+			mutAgent.WaitOne ();
+			agentMessageQueue.Enqueue (msg);  
+			mutAgent.ReleaseMutex ();
+		}
 
-		public byte[] PopAgentState() {
+		public byte[] PopAgentState ()
+		{
 			byte[] received = null;
-			if( agentMessageQueue.Count > 0 ) {
-				received = agentMessageQueue.Dequeue();
+
+			mutAgent.WaitOne ();
+			if (agentMessageQueue.Count > 0) {
+				received = agentMessageQueue.Dequeue ();
 			}
+			mutAgent.ReleaseMutex ();
 
 			return received;
 		}
-			
+
 		void OnApplicationQuit ()
 		{
 			wssv.Stop ();
