+++ unity-sample-environment/Assets/Scripts/SceneController2.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b45c0a9fd5273415f8356330cdb17218
-timeCreated: 1472283854
+guid: 32da652a008ef4f31b40ee344dfafcf6
+timeCreated: 1472959138
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
